backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
httpx
openpyxl
pydantic[email]
pydantic-settings
essentia-tensorflow
